changed size_t in strncmp, strlcpy, strlcat, memcmp, bzero, memcpy, memmove, memset and memchr..... Neu strlcpy und strlcat..... makefile changed to find all files automaticly

+++ libft/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: rbetz <rbetz@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/24 15:10:33 by rbetz             #+#    #+#             */
-/*   Updated: 2022/03/25 12:16:48 by rbetz            ###   ########.fr       */
+/*   Updated: 2022/03/29 17:28:28 by rbetz            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	ft_memset(s, '\0', n);
 }
